requirements.txt
aenum==3.1.0
asgiref==3.5.0
asteval==0.9.23
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
backoff==1.11.1
beautifulsoup4==4.10.0
bintrees==2.2.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
configparser==5.0.2
coverage==6.3.2
cycler==0.11.0
dependency-injector==4.39.1
deprecation==2.1.0
distlib==0.3.4
Django==3.2.12
drawtree==0.1.2
filelock==3.6.0
Flask==2.0.3
fonttools==4.31.2
funcsigs==1.0.2
future==0.18.2
gs-quant==0.8.332
idna==3.2
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.1.1
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
llvmlite==0.36.0
lmfit==1.0.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
more-itertools==8.9.0
msgpack==1.0.2
nest-asyncio==1.5.1
numba==0.53.1
numpy==1.20.3
packaging==21.0
pandas==1.2.4
patsy==0.5.1
Pillow==9.1.0
platformdirs==2.5.1
plotly==5.7.0
pluggy==1.0.0
py==1.11.0
pydash==5.0.2
pylint==2.12.2
pyparsing==2.4.7
pytest==7.1.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
statsmodels==0.12.2
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
tqdm==4.62.2
typed-ast==1.5.2
typing-extensions==4.1.1
uncertainties==3.1.6
urllib3==1.26.6
virtualenv==20.14.0
websockets==10.0
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
